Voodoo sound: introducing usage of the ReTune hardware parametric EQ
This first extension allow usage of the parametric hardware EQ in order to reduce harshness found in high frequencies.
With the current setup, high frequencies around 20kHz create an associated white noise and tend to make highs fatiguing and unclean.
I don't know if it's a limit of the oversampling abilities of the codec, but the method employed here is surprisingly effective.

The only drawback is that alter (reduce) the high frequency response for very high frequencies. It can slightly affect the brightness and it's probably not recommended to use it if your headphone already lack brightness.
However it's tuned to mostly removes noise artifacts and frequencies which are not reproduced accurately in the default codec config.

This anti-alias filter can be tweaked live through sysfs as simple user.
filename: /sys/devices/virtual/voodoo_sound/parametric_equalizer/anti_alias_filter

0: anti-alias filter disabled
1: anti-alias filter enabled, simple mode. 1 high-shelf
2: anti-alias filter enabled, advanced mode. 1 parametric EQ band + 1 high-shelf

For novice listeners, this filter's effect can be difficult to identify.
Testing with udial.wav make a very clear demonstration of its efficiency ;)